#34 refactor : 뷰 로직, 비즈니스 로직 분리
