+++ src/pages/statusPage/components/ReservationList.tsx
@@ -11,75 +11,36 @@ import {
   canceledStrings,
   pastStrings,
 } from '../utils/constants';
-
-interface StringSets {
-  upcomings: string[];
-  cancels: string[];
-  pasts: string[];
-}
+import { returnNoResult, classifyListItems } from '../utils/helpers';
 
 function ReservationList({ hotelList }: { hotelList: any[] }) {
   const location = useLocation();
 
-  const returnNoResult = (locationString: string, stringSets: StringSets) => {
-    const { upcomings, cancels, pasts } = stringSets;
-    switch (locationString) {
-      case '/bookings':
-      case '/bookings/upcoming':
-        return upcomings.map((upcomingString: string, index: number) => (
-          <PStyled key={`${upcomingString[0]}_${index}`}>
-            {upcomingString}
-          </PStyled>
-        ));
-      case '/bookings/canceled':
-        return cancels.map((canceledString: string, index: number) => (
-          <PStyled key={`${canceledString[0]}_${index}`}>
-            {canceledString}
-          </PStyled>
-        ));
-      case '/bookings/past':
-        return pasts.map((pastString: string, index: number) => (
-          <PStyled key={`${pastString[0]}_${index}`}>{pastString}</PStyled>
-        ));
-      default:
-        return '오류가 발생했습니다';
-    }
-  };
+  const noResults = returnNoResult(location.pathname, {
+    upcomings: upcomingStrings,
+    cancels: canceledStrings,
+    pasts: pastStrings,
+  }).map((resultString: string, index: number) => (
+    <PStyled key={`${resultString[0]}_${index}`}>{resultString}</PStyled>
+  ));
 
-  const classifyListItems = () => {
-    let result = [];
-    const today = new Date();
-    switch (location.pathname) {
-      case '/bookings/canceled':
-        result = hotelList.filter((listItem: any) => listItem[3].canceled);
-        break;
-      case '/bookings/past':
-        result = hotelList.filter(
-          (listItem: any) =>
-            isBefore(new Date(listItem[1].checkout), today) || listItem[4].past,
-        );
-        break;
-      default:
-        result = hotelList.filter(
-          (listItem: any) => !listItem[3].canceled && !listItem[4].past,
-        );
-        break;
-    }
-    return result;
-  };
+  const filterCondition = (dateString: string, todayDate: Date) =>
+    isBefore(new Date(dateString), todayDate);
 
-  const hotelsInfo = classifyListItems().map(
-    (reservationInfo: any[], index: number) => {
-      return (
-        <DetailedHotelInfo
-          hotelName={reservationInfo[0].hotelname}
-          occupancy={reservationInfo[2].person}
-          bookingDates={reservationInfo[1].date}
-          key={`${reservationInfo[0].hotelname}_${index}`}
-        />
-      );
-    },
-  );
+  const userReservationInfo = classifyListItems(
+    hotelList,
+    location.pathname,
+    filterCondition,
+  ).map((reservationInfo: any[], index: number) => {
+    return (
+      <DetailedHotelInfo
+        hotelName={reservationInfo[0].hotelname}
+        occupancy={reservationInfo[2].person}
+        bookingDates={reservationInfo[1].date}
+        key={`${reservationInfo[0].hotelname}_${index}`}
+      />
+    );
+  });
 
   return (
     <ReservationListBlock>
@@ -89,14 +50,10 @@ function ReservationList({ hotelList }: { hotelList: any[] }) {
             <DivStyled>
               <AiOutlineCloseCircle />
             </DivStyled>
-            {returnNoResult(location.pathname, {
-              upcomings: upcomingStrings,
-              cancels: canceledStrings,
-              pasts: pastStrings,
-            })}
+            {noResults}
           </NoResultBox>
         )}
-        {(hotelList || (hotelList as any[]).length > 0) && hotelsInfo}
+        {(hotelList || (hotelList as any[]).length > 0) && userReservationInfo}
       </>
     </ReservationListBlock>
   );
